+++ src/components/PostingCard.js
@@ -23,7 +23,7 @@ const PostingCard = () => {
 const PostingContainer = styled.img`
   width: 400px;
   height: 300px;
-  margin: 10px auto;
+  margin: 20px auto;
 
   border-radius: 20px;
   box-shadow: 5px 5px 10px gray;
