+++ src/pages/main/Main.js
@@ -1,11 +1,14 @@
 import React from "react";
 import styled from "styled-components";
-import Header from "../components/Header";
-import PostingCard from "../components/PostingCard";
+import Header from "../../components/Header";
+import PostingCard from "../../components/PostingCard";
+import { RESP } from "../../data/response";
 
 const Main = () => {
+	console.log(RESP.QEUSTION.result);
+
   return (
-    <>
+    <MainWrap>
       <Header />
       <MainContainer>
         <ListHeader>
@@ -18,40 +21,29 @@ const Main = () => {
           })}
         </ListContainer>
       </MainContainer>
-    </>
+    </MainWrap>
   );
 };
+
+const MainWrap = styled.div`
+	width: 99.655vw;
+	height: 99.655vh;
+	display: flex;
+	flex-direction: column;
+`;
 
 const MainContainer = styled.div`
   background-color: aliceblue;
 
-  width: 30rem;
-  height: 80vh;
-  margin: 30px auto;
+  width: 570px;
+  height: 825px;
+  margin: auto;
   text-align: center;
-
-  // scroll
-  /* overflow-y: scroll;
-	
-	&::-webkit-scrollbar {
-		width: 10px;
-	}
-	&::-webkit-scrollbar-track {
-	  box-shadow: inset 0 0 5px grey; 
-	  border-radius: 10px;
-	}
-	&::-webkit-scrollbar-thumb {
-		background: #888;
-		border-radius: 5px;
-	}
-	&::-webkit-scrollbar-thumb:hover {
-		background: #333;
-		border-radius: 5px;
-	} */
 `;
 
 const ListHeader = styled.div`
-  width: 26rem;
+  width: 80%;
+	height: 125px;
   margin: 10px auto;
 
   display: flex;
@@ -67,6 +59,26 @@ const ListHeader = styled.div`
 
 const ListContainer = styled.div`
   background-color: lightgreen;
+	height: 700px;
+	
+  // scroll
+  overflow-y: scroll;
+	
+	&::-webkit-scrollbar {
+		width: 10px;
+	}
+	&::-webkit-scrollbar-track {
+	  box-shadow: inset 0 0 5px grey; 
+	  border-radius: 10px;
+	}
+	&::-webkit-scrollbar-thumb {
+		background: #888;
+		border-radius: 5px;
+	}
+	&::-webkit-scrollbar-thumb:hover {
+		background: #333;
+		border-radius: 5px;
+	}
 `;
 
 export default Main;
