+++ src/pages/signup/SignUp.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../components/Header";
+import Header from "../../components/Header";
 import "./SignUp.css";
 
 const SignUp = () => {
